+++ frontend/src/components/VariantDetails/VariantFreqs.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { isVariantMt } from '@/api/utils'
+import { isVariantMt } from '@/lib/utils'
 
 import VariantDetailsFreqsAutosomal from '@/components/VariantDetails/FreqsAutosomal.vue'
 import VariantDetailsFreqsMitochondrial from '@/components/VariantDetails/FreqsMitochondrial.vue'
