+++ frontend/src/components/__tests__/AcmgRating.spec.ts
@@ -11,7 +11,7 @@ import { createVuetify } from 'vuetify'
 import * as components from 'vuetify/components'
 import * as directives from 'vuetify/directives'
 
-import { MultiSourceAcmgCriteriaState, StateSource, AcmgCriteria, Presence } from '@/api/acmgSeqVar'
+import { MultiSourceAcmgCriteriaState, StateSource, AcmgCriteria, Presence } from '@/lib/acmgSeqVar'
 import AcmgRating from '@/components/VariantDetails/AcmgRating.vue'
 
 const vuetify = createVuetify({
